+++ src/javascript/config.js
@@ -12,173 +12,85 @@ Paris.config = {
     "indexes": {
       "global": "recette_ParisFront",
       "persons": "recette_Elus",
-      "equipments": "Service_production",
-      "dev_mes_lieux": "dev_MesLieux",
+      "equipments": "Service_recette",
     },
     "main_facets": ["Bibliothèques", "Mairies d'arrondissement", "Principaux parcs, jardins et squares", "Piscines", "Tennis"],
     "main_facets_icon": [
       {
-        facet: "Piscines",
-        icon: "swimming-pool"
+        "facet": "Piscines",
+        "icon": "swimming-pool"
       },
       {
-        facet: "Bibliothèques",
-        icon: "library"
+        "facet": "Bibliothèques",
+        "icon": "library"
       },
       {
-        facet: "Médiathèques et ludothèques",
-        icon: "library"
+        "facet": "Médiathèques et ludothèques",
+        "icon": "library"
       },
       {
-        facet: "Principaux parcs, jardins et squares",
-        icon: "park"
+        "facet": "Principaux parcs, jardins et squares",
+        "icon": "park"
       },
       {
-        facet: "Jardins, mails, promenades",
-        icon: "park"
+        "facet": "Jardins, mails, promenades",
+        "icon": "park"
       },
       {
-        facet: "Squares",
-        icon: "park"
+        "facet": "Squares",
+        "icon": "park"
       },
       {
-        facet: "Marchés alimentaires et spécialisés",
-        icon: "market"
+        "facet": "Marchés alimentaires et spécialisés",
+        "icon": "market"
       },
       {
-        facet: "Marchés spécialisés",
-        icon: "market"
+        "facet": "Marchés spécialisés",
+        "icon": "market"
       },
       {
-        facet: "Ateliers Beaux-Arts",
-        icon: "art-workshop"
+        "facet": "Ateliers Beaux-Arts",
+        "icon": "art-workshop"
       },
       {
-        facet: "Centres Paris Anim",
-        icon: "animation"
+        "facet": "Centres Paris Anim",
+        "icon": "animation"
       },
       {
-        facet: "Conservatoires municipaux",
-        icon: "conservatoire"
+        "facet": "Conservatoires municipaux",
+        "icon": "conservatoire"
       },
       {
-        facet: "Mairies d'arrondissement",
-        icon: "city-hall"
+        "facet": "Mairies d'arrondissement",
+        "icon": "city-hall"
       },
       {
-        facet: "Hôtel de Ville",
-        icon: "city-hall"
+        "facet": "Hôtel de Ville",
+        "icon": "city-hall"
       },
       {
-        facet: "Maisons des associations",
-        icon: "associations"
+        "facet": "Maisons des associations",
+        "icon": "associations"
       },
       {
-        facet: "Point Paris Emploi",
-        icon: "jobs"
+        "facet": "Point Paris Emploi",
+        "icon": "jobs"
       },
       {
-        facet: "Musées municipaux",
-        icon: "museum"
+        "facet": "Musées municipaux",
+        "icon": "museum"
       },
       {
-        facet: "Tennis",
-        icon: "tennis"
+        "facet": "Tennis",
+        "icon": "tennis"
       },
       {
-        facet: "Théàtres de la Ville et établissements soutenus",
-        icon: "culture"
+        "facet": "Théàtres de la Ville et établissements soutenus",
+        "icon": "culture"
       },
       {
-        facet: "Théàtres pour enfants",
-        icon: "culture"
-      }
-    ],
-    "secondary_filters": [
-      {
-        "title": "Ouverture",
-        "type": "checkbox",
-        "id": "is_open",
-        "label": "Ouvert maintenant"
-      },
-      {
-        "title": "Accessibilité",
-        "type": "select",
-        "id": "accessibility",
-        "values": [
-          {
-            "id": "all",
-            "label": "Sans handicap"
-          },
-          {
-            "id": "accessible aux personnes à mobilité réduite",
-            "label": "Handicap moteur"
-          },
-          {
-            "id": "accessible aux non-voyants et malvoyants",
-            "label": "Handicap visuel"
-          },
-          {
-            "id": "accessible aux sourds et malentendants",
-            "label": "Handicap auditif"
-          }
-        ]
-      },
-      {
-        "title": "Taille du bassin",
-        "type": "select",
-        "id": "pool_length",
-        "linked_filter": "Piscines",
-        "values": [
-          {
-            "id": "all",
-            "label": "Sans préférence"
-          },
-          {
-            "id": "25m",
-            "label": "25 m"
-          },
-          {
-            "id": "33m",
-            "label": "33 m"
-          },
-          {
-            "id": "50m",
-            "label": "50 m"
-          }
-        ]
-      },
-      {
-        "title": "Section",
-        "type": "select",
-        "id": "sections",
-        "linked_filter": "Bibliothèques",
-        "values": [
-          {
-            "id": "all",
-            "label": "Tous"
-          },
-          {
-            "id": "adulte",
-            "label": "Adulte"
-          },
-          {
-            "id": "bibliothèque spécialisée",
-            "label": "Bibliothèques spécialisées"
-          },
-          {
-            "id": "discothèque",
-            "label": "Discothèques"
-          },
-          {
-            "id": "jeunesse",
-            "label": "Jeunesse"
-          },
-          {
-            "id": "vidéothèque",
-            "label": "Vidéothèques"
-          }
-        ]
+        "facet": "Théàtres pour enfants",
+        "icon": "culture"
       }
     ],
     "url": {
@@ -209,11 +121,13 @@ Paris.config = {
   "captcha": {
     "key": "6Lf2DQoTAAAAAKmk3wEFCZuK3FqG00SlQM3o6Yvp"
   },
-  "mapbox": {
+  "leaflet": {
     "accessToken": "pk.eyJ1IjoicGFyaXNudW1lcmlxdWUiLCJhIjoiY2loZG1vMnYyMDAzNnY0a3FvNG1nNG55biJ9.MP1qcHkEecFGqSTs9gg7cw",
     "defaultCenter": {"lat": 48.856578, "lon": 2.351828},
-    "defaultBounds": [[48.902157, 2.46976], [48.815575, 2.224122]],
-    "styleLayer": "mapbox://styles/parisnumerique/cis1rkqoj000khinpthppoaqd"
+    "minZoom": 11,
+    "maxBounds": [[49.2,2.56], [48.76,2.11]],
+    "tileLayer": "http://filer.paris.fr/leaflet/paris2/{z}/{x}/{y}.png",
+    "tileLayerAttribution": "&copy; <a href=\"http://osm.org/copyright\">OpenStreetMap</a> contributors"
   },
   "search": {
     "paris_coordinates" : {