+++ src/javascript/locales.js
@@ -36,7 +36,6 @@ Paris.i18n.locales = {
       "reset_filters": "Réinitialiser",
       "more_filters": "Afficher la liste complète",
       "more_filters_nb": "Et {0} de plus",
-      "apply_filters": "Appliquer",
       "search_filter": "Rechercher un filtre...",
       "around_me": "Autour de moi",
       "equipements": "Équipements",
@@ -70,6 +69,9 @@ Paris.i18n.locales = {
     "active": "actif",
     "inactive": "inactif",
     "new_window": "nouvelle fenêtre",
-    "close_nav": "fermer le menu \"{0}\""
+    "close_nav": "fermer le menu \"{0}\"",
+    "leaflet": {
+      "sleep_button": "désactiver la carte"
+    }
   }
 };